+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/util/RealmManager.java
@@ -186,4 +186,11 @@ public class RealmManager {
         realm.update(rep);
         return this;
     }
+
+    public RealmManager defaultAudValueForAccessToken(String defaultAudValueForAccessToken) {
+        RealmRepresentation realmRepresentation = realm.toRepresentation();
+        realmRepresentation.setDefaultAudValueForAccessToken(defaultAudValueForAccessToken);
+        realm.update(realmRepresentation);
+        return this;
+    }
 }